Int becomes bitvector
Leon now matches Scala semantics of Int as 32 bits bit-vectors.
This commits modifies the semantics of IntLiteral to be treated
as 32 bits integer everywhere (solver, evaluator, ...).

Introduces a new literal type, InfiniteIntegerLiteral, representing
a natural integer. The front-end maps the use of BigInt to these
new trees, and the solver properly handles them as mathematical
integers.

The behaviour of many regression tests changes due to this new
semantics. In particular many of them now timeout because they
are no longer proving properties over mathematical integers. This
commit updates the tests to reflect this new semantics.

+++ src/test/resources/regression/verification/xlang/valid/Nested7.scala
@@ -2,10 +2,10 @@
 
 object Nested2 {
 
-  def foo(a: Int): Int = {
+  def foo(a: BigInt): BigInt = {
     require(a >= 0)
     val b = a + 2
-    def rec1(c: Int): Int = {
+    def rec1(c: BigInt): BigInt = {
       require(c >= 0)
       b + c + bar(a) + bar(b) + bar(c)
     }
@@ -13,7 +13,7 @@ object Nested2 {
   } ensuring(_ > 0)
 
 
-  def bar(x: Int): Int = {
+  def bar(x: BigInt): BigInt = {
     require(x >= 0)
     x
   } ensuring(_ >= 0)
